megaplan: outcome-astrid-has-a-single-20260602-1656 execute
