+++ astrid/packs/reigh/executors/open_in_reigh/run.py
@@ -31,7 +31,7 @@ from pathlib import Path
 
 from astrid.contracts.errors import AstridError
 from astrid.packs._canonical_entrypoint import run_pack_main
-from astrid.timeline import Timeline
+from astrid.core.timeline import Timeline
 
 DEFAULT_REIGH_APP = Path("/Users/peteromalley/Documents/reigh-workspace/reigh-app")
 PROBE_DIRS = ("public/timelines", "public/demos", "timelines", "demos")
