+++ astrid/packs/training/executors/pool_merge/run.py
@@ -13,7 +13,7 @@ import json
 from pathlib import Path
 from typing import Any, Sequence
 
-from astrid import timeline
+from astrid.core import timeline
 from astrid.audit import register_outputs
 from astrid.core.element import catalog as effects_catalog
 from astrid.core.util.time import utc_now_seconds
